requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.6.1
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
mysql-connector-python==9.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
PyMySQL==1.1.1
pytest==8.3.2
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.2.2